requirements.txt
edge-tts
moviepy==1.0.3
requests
Pillow
numpy
